# docs/openapi-v1.yaml
openapi: 3.0.3
info:
  title: WrkPlan API
  version: 1.0.0
  description: API contract for Laravel-to-.NET migration.
servers:
  - url: /api/v1
components:
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: wrkplan_session
...
security:
  - sessionAuth: []
paths:
...
  /admin/tickets/{supportTicket}/reply:
    post:
      summary: Reply to support ticket
      parameters:
        - in: path
          name: supportTicket
          required: true
          schema: { type: integer }
      responses:
        '200': { description: Replied }
  /admin/tickets/{supportTicket}/status:
    patch:
      summary: Update support ticket status
      parameters:
        - in: path
          name: supportTicket
          required: true
          schema: { type: integer }
      responses:
        '200': { description: Updated }
  /admin/tickets/{supportTicket}/assign:
    patch:
      summary: Assign support ticket
      parameters:
        - in: path
          name: supportTicket
          required: true
          schema: { type: integer }
      responses:
        '200': { description: Updated }
...
  /customer/tickets/{id}/reply:
    post:
      summary: Reply to customer ticket
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        '200': { description: Replied }
...